Rewritten transition backend
This allows for slides to have different transitions without breaking visually, almost allows for actually changing the transition duration, and makes a distinction between "in" keyframes (to be applied to the new slide) and "out" keyframes (to be applied to the old slide). If no out keyframes are specified, the old slide gets the "in" keyframes in reverse.

Next, rewriting the transition editor.