Allow forward/reverse buttons to traverse the whole show
Also, fix the UI a bit.

Known issue: having the same deck in a show more than once will cause strange behavior - at the end of the second instance of the deck, pressing "forward" again will go back to the next slide after the *first* instance of the deck. A fix for this issue is currently not planned, but may happen eventually.